JS: Make chunks an array of RefValue
Towards #1112
