+++ js/src/collection.js
@@ -1,13 +1,13 @@
-  // @flow
+// @flow
 
-import Ref from './ref.js';
+import RefValue from './ref-value.js';
 import type {Sequence} from './sequence.js'; // eslint-disable-line no-unused-vars
 import {isPrimitive} from './primitives.js';
-import {MetaTuple} from './meta-sequence.js';
-import {Type} from './type.js';
+import type {MetaTuple} from './meta-sequence.js';
+import type {Type} from './type.js';
 import {ValueBase} from './value.js';
 
-export class Collection<S:Sequence> extends ValueBase {
+export class Collection<S: Sequence> extends ValueBase {
   sequence: S;
 
   constructor(type: Type, sequence: S) {
@@ -19,10 +19,11 @@ export class Collection<S:Sequence> extends ValueBase {
     return !this.sequence.isMeta && this.sequence.items.length === 0;
   }
 
-  get chunks(): Array<Ref> {
+  get chunks(): Array<RefValue> {
     const chunks = [];
-    const addChunks = this.sequence.isMeta ? (mt:MetaTuple) => {
-      chunks.push(mt.ref);
+    const addChunks = this.sequence.isMeta ? (mt: MetaTuple) => {
+      // TODO: https://github.com/attic-labs/noms/issues/1106
+      chunks.push(new RefValue(mt.ref));
     } : (v) => {
       if (!isPrimitive(v)) {
         chunks.push(...v.chunks);
