+++ js/src/get-ref.js
@@ -1,10 +1,10 @@
 // @flow
 
-import Chunk from './chunk.js';
-import Ref from './ref.js';
+import type Chunk from './chunk.js';
+import type Ref from './ref.js';
 import type DataStore from './data-store.js';
 import {notNull} from './assert.js';
-import {Type} from './type.js';
+import type {Type} from './type.js';
 
 type encodeFn = (v: any, t: Type, ds: ?DataStore) => Chunk;
 let encodeNomsValue: ?encodeFn = null;
