+++ js/src/meta-sequence.js
@@ -5,8 +5,9 @@ import {default as Ref, sha1Size} from './ref.js';
 import type {BoundaryChecker, makeChunkFn} from './sequence-chunker.js';
 import type DataStore from './data-store.js';
 import type {valueOrPrimitive} from './value.js'; // eslint-disable-line no-unused-vars
-import {Collection} from './collection.js';
-import {CompoundDesc, makeCompoundType, makePrimitiveType, Type} from './type.js';
+import type {Collection} from './collection.js';
+import {CompoundDesc, makeCompoundType, makePrimitiveType} from './type.js';
+import type {Type} from './type.js';
 import {IndexedSequence} from './indexed-sequence.js';
 import {invariant} from './assert.js';
 import {Kind} from './noms-kind.js';
