+++ js/src/noms.js
@@ -18,6 +18,7 @@ export {lookupPackage, Package, readPackage, registerPackage} from './package.js
 export {newList, ListLeafSequence, NomsList} from './list.js';
 export {newMap, NomsMap, MapLeafSequence} from './map.js';
 export {newSet, NomsSet, SetLeafSequence} from './set.js';
+export {IndexedSequence} from './indexed-sequence.js';
 export {OrderedMetaSequence, IndexedMetaSequence} from './meta-sequence.js';
 export {SPLICE_AT, SPLICE_REMOVED, SPLICE_ADDED, SPLICE_FROM} from './edit-distance.js';
 export {
