+++ js/src/value.js
@@ -1,15 +1,16 @@
 // @flow
 
-import Ref from './ref.js';
+import type Ref from './ref.js';
 import type {primitive} from './primitives.js';
 import {ensureRef} from './get-ref.js';
-import {Type} from './type.js';
+import type {Type} from './type.js';
+import type RefValue from './ref-value.js';
 
 export interface Value {
   ref: Ref;
   equals(other: Value): boolean;
   less(other: Value): boolean;
-  chunks: Array<Ref>;
+  chunks: Array<RefValue>;
   type: Type;
 }
